requirements.txt
PySide6
mysql-connector-python
openpyxl
bcrypt>=4.0.1
